linux/modules: Move package under datapath/linux
Kernel modules manager is meant to work with Linux kernel modules, so
it belongs to the datapath/linux folder.

CODEOWNERS file is updated accordingly.
